refactor(snippets): 기존 스니펫 정보 마이그레이션, 지원중단할 예정
