+++ packages/core/src/snippets/request-parts-snippet.ts
@@ -2,7 +2,7 @@ import { isEmpty } from "es-toolkit/compat";
 
 import { format } from "@/utils/format";
 
-import type { PartDescriptor } from "@/types";
+import type { PartDescriptor } from "@/descriptors";
 
 /**
  * request-parts
