+++ packages/core/src/types/doc-options.type.ts
@@ -1,5 +1,3 @@
-import type { FieldDescriptor } from "./descriptors.type";
-
 /** 문서 포맷 타입 */
 export type DocumentFormat = "adoc" | "md";
 
@@ -11,18 +9,4 @@ export type NRestDocsConfig = {
     format: DocumentFormat;
     /** strict 모드 시, 문서 정의와 실제 요청/응답 불일치 시 테스트 실패 */
     strict: boolean;
-};
-
-/** 테스트 코드에서 문서화 시 제공하는 옵션 */
-export type DocOptions = {
-    /** API에 대한 간단한 설명 */
-    description?: string;
-
-    /** 요청 JSON 필드 정보 */
-    requestFields?: FieldDescriptor[];
-
-    /** 응답 JSON 필드 정보 */
-    responseFields?: FieldDescriptor[];
-
-    // pathParams, queryParams, headers
 };
