+++ packages/core/tsconfig.json
@@ -19,5 +19,6 @@
             "@/*": ["src/*"]
         }
     },
-    "include": ["src/**/*.ts", "vitest.config.ts"] // 컴파일에 포함할 파일
+    "include": ["src/**/*.ts", "vitest.config.ts"], // 컴파일에 포함할 파일
+    "exclude": ["node_modules", "dist", "**/*.spec.ts"]
 }
